refactor(ui): rename inline state helpers
