+++ apps/ui/src/app/views/inline_state_panel_models.ts
